+++ src/lib.rs
@@ -146,9 +146,13 @@ impl Orthanc {
 
     ////////// STUDY //////////
 
-    pub fn list_studies(&self) -> Result<Table> {
+    pub fn list_studies(&self, patient_id: Option<&str>) -> Result<Table> {
+        let mut studies = self.client.studies_expanded()?;
+        if let Some(pid) = patient_id {
+            studies.retain(|s| s.parent_id().unwrap() == pid);
+        };
         Ok(create_list_table(
-            self.client.studies_expanded()?,
+            studies,
             &STUDIES_LIST_HEADER,
             &STUDIES_LIST_DICOM_TAGS,
         ))
@@ -210,9 +214,13 @@ impl Orthanc {
 
     ////////// SERIES //////////
 
-    pub fn list_series(&self) -> Result<Table> {
+    pub fn list_series(&self, study_id: Option<&str>) -> Result<Table> {
+        let mut series = self.client.series_expanded()?;
+        if let Some(pid) = study_id {
+            series.retain(|s| s.parent_id().unwrap() == pid);
+        };
         Ok(create_list_table(
-            self.client.series_expanded()?,
+            series,
             &SERIES_LIST_HEADER,
             &SERIES_LIST_DICOM_TAGS,
         ))
@@ -274,9 +282,13 @@ impl Orthanc {
 
     ////////// INSTANCE //////////
 
-    pub fn list_instances(&self) -> Result<Table> {
+    pub fn list_instances(&self, series_id: Option<&str>) -> Result<Table> {
+        let mut instances = self.client.instances_expanded()?;
+        if let Some(pid) = series_id {
+            instances.retain(|s| s.parent_id().unwrap() == pid);
+        };
         Ok(create_list_table(
-            self.client.instances_expanded()?,
+            instances,
             &INSTANCES_LIST_HEADER,
             &INSTANCES_LIST_DICOM_TAGS,
         ))